ANDAUTH-30: Rewrite low-level usable of HTTP requests by Retrofit * extract SyncData
