+++ app/src/main/java/org/xwiki/android/authenticator/rest/SyncData.java
@@ -1,35 +1,38 @@
 package org.xwiki.android.authenticator.rest;
 
-import org.xwiki.android.authenticator.bean.XWikiUser;
+import org.xwiki.android.authenticator.bean.XWikiUserFull;
 
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 public class SyncData {
     //the users which have been modified from the last sync time. mainly used for updating and adding..
-    List<XWikiUser> updateUserList;
-    //all the users in the server or all the users of the selected groups used for deleting.
-    HashSet<String> allIdSet;
+    private List<XWikiUserFull> updateUserList = new ArrayList<>();
 
-    public SyncData() {
-        updateUserList = new ArrayList<>();
-        allIdSet = new HashSet<>();
-    }
+    private Set<String> additionalIds = new HashSet<>();
 
-    public List<XWikiUser> getUpdateUserList() {
+    public List<XWikiUserFull> getUpdateUserList() {
         return updateUserList;
     }
 
-    public HashSet<String> getAllIdSet() {
-        return allIdSet;
+    public Set<String> getAllIdSet() {
+        Set<String> idsSet = new HashSet<>(additionalIds);
+        for (XWikiUserFull user : getUpdateUserList()) {
+            idsSet.add(user.id);
+        }
+        return idsSet;
+    }
+
+    public void addAdditionalId(String id) {
+        additionalIds.add(id);
     }
 
     @Override
     public String toString() {
         return "SyncData{" +
-                "updateUserList=" + updateUserList +
-                ", allIdSet=" + allIdSet +
-                '}';
+            "updateUserList=" + updateUserList +
+            '}';
     }
 }
